refactor(direct-control): extract setup read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only setup phase, setup snapshot/map-row, setup map-row read, and setup map-row visibility result type ownership from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/setup/reads.ts while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not change setup snapshot/map-row source strings, map-script validation, setup map-row refresh behavior, setup lifecycle mutation behavior, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Setup prepare/start/run lifecycle input/result types remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- setup-and-lifecycle.test.ts public-api.test.ts restart-lifecycle.test.ts
- rg -n "setup|single-player|startPrepared|prepareCiv7" packages/cli/test/commands packages/cli/src/commands
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, setup lifecycle behavior, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
